+++ src/main/java/com/example/msa/user/service/user/adapter/in/web/dto/request/UserSignUpRequest.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserSignUpRequest {
-    private String id;
+    private String userId;
     private String pw;
     private String role;
 }
