+++ src/main/java/com/example/msa/user/service/user/application/service/UserService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.msa.user.service.user.adapter.out.db.PersistenceAdapter;
-import com.example.msa.user.service.user.application.model.domain.User;
+import com.example.msa.user.service.user.domain.model.User;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,29 +24,29 @@ public class UserService {
     /**
      * <b> 유저 회원가입 로직 수행 </b>
      *
-     * @param id 아이디
+     * @param userId 사용자가 작성한 아이디
      * @param pw 비밀번호
      * @param role 역할
      * @return 회원가입 성공 여부
      */
-    public boolean signUp(String id, String pw, String role) {
-        User user = User.createUser(id, pw, role);
-        persistenceAdapter.saveUser(user);
-        return true;
+    public boolean signUp(String userId, String pw, String role) {
+        return persistenceAdapter.create(userId, pw, role);
     }
 
     /**
      * <b> 유저 로그인 로직 수행 </b>
      *
+     * @param userId 회원가입 시 아이디
      * @param pw 비밀번호
      * @return 로그인 성공 여부
      */
-    public boolean signIn(String id, String pw) {
-        User user = persistenceAdapter.findById(id);
-        // if (user == null || !user.getPw().equals(pw)) {
-        //     logger.log(Level.WARNING, "로그인 실패 : 잘못된 ID 또는 비밀번호");
-        //     return false;
-        // }
-        return true;
+    public boolean signIn(String userId, String pw) {
+        User user = persistenceAdapter.findByUserId(userId);
+         if (user != null && user.getPw().equals(pw)) {
+             return true;
+         }
+
+        logger.log(Level.WARNING, "로그인 실패 : 잘못된 ID 또는 비밀번호");
+        return false;
     }
 }
